chore: move components to seperate files
