+++ src/app/dashboard/events/[id]/participants/table/delete-dialog.tsx
@@ -1,0 +1,73 @@
+"use client";
+
+import { Trash2, XCircle } from "lucide-react";
+
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from "@/components/ui/alert-dialog";
+import { Button, buttonVariants } from "@/components/ui/button";
+
+export function DeleteParticipantDialog({
+  isQuerying,
+  participantId,
+  deleteParticipant,
+}: {
+  isQuerying: boolean;
+  participantId: number;
+  deleteParticipant: (_participantId: number) => Promise<void>;
+}) {
+  return (
+    <AlertDialog>
+      <AlertDialogTrigger asChild>
+        <Button
+          variant="outline"
+          className="h-12 w-fit px-3 text-red-500"
+          disabled={isQuerying}
+          size="icon"
+        >
+          <Trash2 />
+          Usuń
+        </Button>
+      </AlertDialogTrigger>
+      <AlertDialogContent className="flex flex-col items-center">
+        <AlertDialogHeader className="flex flex-col items-center">
+          <AlertDialogTitle className="flex flex-col items-center self-center">
+            <XCircle strokeWidth={1} stroke={"red"} size={64} />
+            Jesteś pewny?
+          </AlertDialogTitle>
+          <AlertDialogDescription className="text-foreground text-center text-pretty">
+            Na pewno chcesz usunąć tego uczestnika? Tej operacji nie będzie
+            można cofnąć.
+          </AlertDialogDescription>
+        </AlertDialogHeader>
+        <AlertDialogFooter className="flex gap-x-4">
+          <AlertDialogAction
+            onClick={async () => {
+              await deleteParticipant(participantId);
+            }}
+            className={buttonVariants({
+              variant: "destructive",
+            })}
+          >
+            Usuń
+          </AlertDialogAction>
+          <AlertDialogCancel
+            className={buttonVariants({
+              variant: "outline",
+            })}
+          >
+            Anuluj
+          </AlertDialogCancel>
+        </AlertDialogFooter>
+      </AlertDialogContent>
+    </AlertDialog>
+  );
+}
