+++ src/app/dashboard/events/[id]/participants/table/table-row-form.tsx
@@ -23,11 +23,9 @@ import type { Block } from "@/types/blocks";
 import type { FlattenedParticipant } from "@/types/participant";
 
 import { updateParticipant } from "../actions";
-import {
-  DeleteParticipantDialog,
-  EditParticipantButton,
-} from "./action-components";
+import { DeleteParticipantDialog } from "./delete-dialog";
 import { DownloadAttributeFileButton } from "./download-file-attribute-button";
+import { EditParticipantButton } from "./edit-button";
 
 interface TableRowFormProps {
   row: Row<FlattenedParticipant>;
